fix issue with xkcd retrieval, some unnoticed bugs

+++ start.sh
@@ -2,20 +2,26 @@
 
 cd "$(dirname "$0")"
 
+touch resources/.log
+
+if sudo docker ps | grep -q "owen-bot"; then
+	sudo docker logs owen-bot &>> resources/.log
+	sudo docker kill "owen-bot"
+fi
+
 if [ "$1" == "-u" ]; then
 	cp resources/config.json resources/config.json.backup
 	git stash
 	git pull
 	cp resources/config.json.backup resources/config.json
+
+	$1="-b"
 fi
 
-touch .log
-sudo docker logs owen-bot >> .log
-
-if sudo docker ps | grep -q "owen-bot"; then
-	sudo docker kill "owen-bot"
+if [ "$1" == "-b" ]; then
+	yes | sudo docker system prune
+	sudo docker build -t "owen-bot" .
 fi
-yes | sudo docker system prune
-sudo docker build -t "owen-bot" .
+
 sudo docker run -d --name "owen-bot" -e TZ=Australia/Melbourne \
 	-v $(pwd)/resources:/owen-bot/resources "owen-bot"
